refactor(schema): extract declaration merge logic
